requirements.txt
numpy
Pillow
torch==1.6
torchvision
tqdm